refactor(#255): extract execution.rs (RunContext, run_parsed_command_with_mode, run_tool)
Move OutputFormat, RunContext, ParsedCommandConfig, combine_output, obtain_output,
render_output, passthrough_raw, record_and_report, run_parsed_command_with_mode,
format_analytics_label, ToolRunConfig, and run_tool from mod.rs into a new
crates/rskim/src/cmd/execution.rs module.

Re-export all pub(crate) items from mod.rs so all callers continue to work
without changes outside cmd/. format_analytics_label tests moved to execution.rs.
Removed now-unused imports (Cow, ParseResult, CommandOutput, CommandRunner, Write)
from mod.rs.

mod.rs drops from 2,070 to 1,032 lines. All 3,640+ tests pass.
